Trend Approval Rate of The Fed
By scraping and analyzing 2 million tweets related to the Federal Reserve posted in the past 7 years, we find the public's confidence in the Fed has been constantly changing. Roughly speaking, the continuing drop in the years before 2015 coincided with the tapering of quantitative easing. The rise between 2015 and 2018 came at a time when the Fed tried to bring monetary policy back to the normal track. And the turning point at the beginning of 2018 coincided well with the ending of Yellen's four year term.